Various fixes

+++ faculty.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 // Recursively calculate the factorial of n
-int factorial(int n)
+long long factorial(int n)
 {
     if (n == 1)
     {
@@ -17,9 +17,9 @@ int factorial(int n)
 }
 
 // Iteratively calculate the factorial of n
-int factorial_iter(int n)
+long long factorial_iter(int n)
 {
-    int result = 1;
+    long long result = 1;
     for (int i = 2; i <= n; i++)
     {
         result *= i;
